query/function/requirements.txt
# Function dependencies, for example:
# package>=version
flask
click<8.0.0
pymysql>=0.9
google.cloud.logging
flask_cors
pytz
Pillow
Werkzeug==2.2.2
numpy
